+++ core/src/main/scala/org/apache/spark/network/BlockDataManager.scala
@@ -22,7 +22,7 @@ import scala.reflect.ClassTag
 import org.apache.spark.TaskContext
 import org.apache.spark.network.buffer.ManagedBuffer
 import org.apache.spark.network.client.StreamCallbackWithID
-import org.apache.spark.network.corruption.Cause
+import org.apache.spark.network.shuffle.checksum.Cause
 import org.apache.spark.storage.{BlockId, StorageLevel}
 
 private[spark]
